# platform/wanwu/docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- | --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n| X-Org-Id      | 组织ID    |\r\n\r\n## HTTP Status\r\n| HTTP Status             | 说明                   |\r\n| ----------------------- | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n| 400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败            |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n| 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n| guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      | 【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n| permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      | 权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n## `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n    \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\": \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n        {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n        {\"perm\": \"permission.role\"}\r\n      ]\r\n    }\r\n  },\r\n  \"msg\": \"操作成功\"\r\n}\r\n```",
...
        "contact": {},
...
    "basePath": "/v1",
    "paths": {
...
        "/appspace/rag": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "获取RAG信息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "ragId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.RagInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "更新RAG基本信息",
                "parameters": [
                    {
                        "description": "更新RAG基本信息的请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RagBrief"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "创建RAG",
                "parameters": [
                    {
                        "description": "创建RAG的请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppBriefConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/request.RagReq"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "删除RAG",
                "parameters": [
                    {
                        "description": "删除RAG的请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RagReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/appspace/rag/config": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "更新RAG配置信息",
                "parameters": [
                    {
                        "description": "更新RAG配置信息的请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RagConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/rag/chat": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "私域 RAG 问答",
                "parameters": [
                    {
                        "description": "RAG问答请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ChatRagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "mp.AppModelParams": {
            "type": "object",
            "properties": {
                "providerOpenAICompatible": {
                    "description": "OpenAI-API-compatible模型配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.AppModelParamsOpenAICompatible"
                        }
                    ]
                },
                "providerYuanjing": {
                    "description": "YuanJing模型配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.AppModelParamsYuanjing"
                        }
                    ]
                }
            }
        },
        "mp.AppModelParamsOpenAICompatible": {
            "type": "object",
            "properties": {
                "llm": {
                    "description": "大语言模型配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_openai_compatible.LLMParams"
                        }
                    ]
                }
            }
        },
        "mp.AppModelParamsYuanjing": {
            "type": "object",
            "properties": {
                "llm": {
                    "description": "大语言模型配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_yuanjing.LLMParams"
                        }
                    ]
                }
            }
        },
...
        "mp_openai_compatible.LLMParams": {
            "type": "object",
            "properties": {
                "frequencyPenalty": {
                    "description": "频率惩罚",
                    "type": "number"
                },
                "frequencyPenaltyEnable": {
                    "description": "频率惩罚(开关)",
                    "type": "boolean"
                },
                "maxTokens": {
                    "description": "最大标记",
                    "type": "integer"
                },
                "maxTokensEnable": {
                    "description": "最大标记(开关)",
                    "type": "boolean"
                },
                "presencePenalty": {
                    "description": "存在惩罚",
                    "type": "number"
                },
                "presencePenaltyEnable": {
                    "description": "存在惩罚(开关)",
                    "type": "boolean"
                },
                "temperature": {
                    "description": "温度",
                    "type": "number"
                },
                "temperatureEnable": {
                    "description": "温度(开关)",
                    "type": "boolean"
                },
                "topP": {
                    "description": "Top P",
                    "type": "number"
                },
                "topPEnable": {
                    "description": "Top P(开关)",
                    "type": "boolean"
                }
            }
        },
...
        "mp_yuanjing.LLMParams": {
            "type": "object",
            "properties": {
                "frequencyPenalty": {
                    "description": "频率惩罚",
                    "type": "number"
                },
                "frequencyPenaltyEnable": {
                    "description": "频率惩罚(开关)",
                    "type": "boolean"
                },
                "maxTokens": {
                    "description": "最大标记",
                    "type": "integer"
                },
                "maxTokensEnable": {
                    "description": "最大标记(开关)",
                    "type": "boolean"
                },
                "presencePenalty": {
                    "description": "存在惩罚",
                    "type": "number"
                },
                "presencePenaltyEnable": {
                    "description": "存在惩罚(开关)",
                    "type": "boolean"
                },
                "temperature": {
                    "description": "温度",
                    "type": "number"
                },
                "temperatureEnable": {
                    "description": "温度(开关)",
                    "type": "boolean"
                },
                "topP": {
                    "description": "Top P",
                    "type": "number"
                },
                "topPEnable": {
                    "description": "Top P(开关)",
                    "type": "boolean"
                }
            }
        },
...
        "request.AppBriefConfig": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "avatar": {
                    "description": "图标",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                }
            }
        },
        "request.AppKnowledgeBase": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "知识库id",
                    "type": "string"
                },
                "name": {
                    "description": "知识库名称(请求非必填)",
                    "type": "string"
                }
            }
        },
        "request.AppKnowledgebaseConfig": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "知识库参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppKnowledgebaseParams"
                        }
                    ]
                },
                "knowledgebases": {
                    "description": "知识库id、名字",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.AppKnowledgeBase"
                    }
                }
            }
        },
        "request.AppKnowledgebaseParams": {
            "type": "object",
            "properties": {
                "maxHistory": {
                    "description": "最长上下文",
                    "type": "integer"
                },
                "maxHistoryEnable": {
                    "description": "最长上下文(开关)",
                    "type": "boolean"
                },
                "threshold": {
                    "description": "过滤阈值",
                    "type": "number"
                },
                "thresholdEnable": {
                    "description": "过滤阈值(开关)",
                    "type": "boolean"
                },
                "topK": {
                    "description": "知识条数",
                    "type": "integer"
                },
                "topKEnable": {
                    "description": "知识条数(开关)",
                    "type": "boolean"
                }
            }
        },
        "request.AppModelConfig": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "模型配置"
                },
                "displayName": {
                    "description": "模型展示名称(请求非必填)",
                    "type": "string"
                },
                "examples": {
                    "description": "仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.AppModelParams"
                        }
                    ]
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelId": {
                    "description": "模型ID",
                    "type": "string"
                },
                "modelType": {
                    "description": "模型类型(llm/embedding/rerank)",
                    "type": "string"
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string"
                }
            }
        },
...
        "request.Avatar": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png",
                    "type": "string"
                },
                "path": {
                    "description": "前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)",
                    "type": "string"
                }
            }
        },
...
        "request.ChatRagRequest": {
            "type": "object",
            "required": [
                "question",
                "ragId"
            ],
            "properties": {
                "question": {
                    "type": "string"
                },
                "ragId": {
                    "type": "string"
                }
            }
        },
...
        "request.RagBrief": {
            "type": "object",
            "required": [
                "name",
                "ragId"
            ],
            "properties": {
                "avatar": {
                    "description": "图标",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "ragId": {
                    "type": "string"
                }
            }
        },
        "request.RagConfig": {
            "type": "object",
            "required": [
                "knowledgeBaseConfig",
                "modelConfig",
                "ragId",
                "rerankConfig"
            ],
            "properties": {
                "knowledgeBaseConfig": {
                    "description": "知识库",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppKnowledgebaseConfig"
                        }
                    ]
                },
                "modelConfig": {
                    "description": "模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppModelConfig"
                        }
                    ]
                },
                "ragId": {
                    "type": "string"
                },
                "rerankConfig": {
                    "description": "Rerank模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppModelConfig"
                        }
                    ]
                }
            }
        },
        "request.RagReq": {
            "type": "object",
            "required": [
                "ragId"
            ],
            "properties": {
                "ragId": {
                    "type": "string"
                }
            }
        },
...
        "response.RagInfo": {
            "type": "object",
            "required": [
                "knowledgeBaseConfig",
                "modelConfig",
                "name",
                "ragId",
                "rerankConfig"
            ],
            "properties": {
                "avatar": {
                    "description": "图标",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "knowledgeBaseConfig": {
                    "description": "知识库",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppKnowledgebaseConfig"
                        }
                    ]
                },
                "modelConfig": {
                    "description": "模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppModelConfig"
                        }
                    ]
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "ragId": {
                    "type": "string"
                },
                "rerankConfig": {
                    "description": "Rerank模型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.AppModelConfig"
                        }
                    ]
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}